API Server errors: Not Leader
Adds a new error type - Not Leader to the api server error types. This
will be used to tell the client that it can not process the raft
application without trying to move the request to the leader.

Additionally, I've moved the error types to a new file to make it a lot
easier to understand the code.
