+++ apiserver/errors/errors.go
@@ -8,176 +8,17 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/go-macaroon-bakery/macaroon-bakery/v3/bakery"
 	"github.com/juju/errors"
 	"github.com/juju/loggo"
-	"github.com/juju/names/v4"
 	"github.com/juju/txn"
-	"gopkg.in/macaroon.v2"
 
 	"github.com/juju/juju/apiserver/params"
 	"github.com/juju/juju/core/leadership"
 	"github.com/juju/juju/core/lease"
-	"github.com/juju/juju/core/network"
 	stateerrors "github.com/juju/juju/state/errors"
 )
 
 var logger = loggo.GetLogger("juju.apiserver.common.errors")
-
-func NotSupportedError(tag names.Tag, operation string) error {
-	return errors.Errorf("entity %q does not support %s", tag, operation)
-}
-
-type noAddressSetError struct {
-	unitTag     names.UnitTag
-	addressName string
-}
-
-func (e *noAddressSetError) Error() string {
-	return fmt.Sprintf("%q has no %s address set", e.unitTag, e.addressName)
-}
-
-func NoAddressSetError(unitTag names.UnitTag, addressName string) error {
-	return &noAddressSetError{unitTag: unitTag, addressName: addressName}
-}
-
-func isNoAddressSetError(err error) bool {
-	_, ok := err.(*noAddressSetError)
-	return ok
-}
-
-type unknownModelError struct {
-	uuid string
-}
-
-func (e *unknownModelError) Error() string {
-	return fmt.Sprintf("unknown model: %q", e.uuid)
-}
-
-func UnknownModelError(uuid string) error {
-	return &unknownModelError{uuid: uuid}
-}
-
-func isUnknownModelError(err error) bool {
-	_, ok := err.(*unknownModelError)
-	return ok
-}
-
-// DischargeRequiredError is the error returned when a macaroon requires
-// discharging to complete authentication.
-type DischargeRequiredError struct {
-	Cause          error
-	LegacyMacaroon *macaroon.Macaroon
-	Macaroon       *bakery.Macaroon
-}
-
-// Error implements the error interface.
-func (e *DischargeRequiredError) Error() string {
-	return e.Cause.Error()
-}
-
-// IsDischargeRequiredError reports whether the cause
-// of the error is a *DischargeRequiredError.
-func IsDischargeRequiredError(err error) bool {
-	_, ok := errors.Cause(err).(*DischargeRequiredError)
-	return ok
-}
-
-// IsUpgradeInProgressError returns true if this error is caused
-// by an upgrade in progress.
-func IsUpgradeInProgressError(err error) bool {
-	if stateerrors.IsUpgradeInProgressError(err) {
-		return true
-	}
-	return errors.Cause(err) == params.UpgradeInProgressError
-}
-
-// UpgradeSeriesValidationError is the error returns when a upgrade-series
-// can not be run because of a validation error.
-type UpgradeSeriesValidationError struct {
-	Cause  error
-	Status string
-}
-
-// Error implements the error interface.
-func (e *UpgradeSeriesValidationError) Error() string {
-	return e.Cause.Error()
-}
-
-// IsUpgradeSeriesValidationError returns true if this error is caused by a
-// upgrade-series validation error.
-func IsUpgradeSeriesValidationError(err error) bool {
-	_, ok := errors.Cause(err).(*UpgradeSeriesValidationError)
-	return ok
-}
-
-// errIncompatibleSeries is a standard error to indicate that the series
-// requested is not compatible with the charm of the application.
-type errIncompatibleSeries struct {
-	seriesList []string
-	series     string
-	charmName  string
-}
-
-func NewErrIncompatibleSeries(seriesList []string, series, charmName string) error {
-	return &errIncompatibleSeries{
-		seriesList: seriesList,
-		series:     series,
-		charmName:  charmName,
-	}
-}
-
-func (e *errIncompatibleSeries) Error() string {
-	return fmt.Sprintf("series %q not supported by charm %q, supported series are: %s",
-		e.series, e.charmName, strings.Join(e.seriesList, ", "))
-}
-
-// IsIncompatibleSeriesError returns if the given error or its cause is
-// errIncompatibleSeries.
-func IsIncompatibleSeriesError(err interface{}) bool {
-	if err == nil {
-		return false
-	}
-	// In case of a wrapped error, check the cause first.
-	value := err
-	cause := errors.Cause(err.(error))
-	if cause != nil {
-		value = cause
-	}
-	_, ok := value.(*errIncompatibleSeries)
-	return ok
-}
-
-// RedirectError is the error returned when a model (previously accessible by
-// the user) has been migrated to a different controller.
-type RedirectError struct {
-	// Servers holds the sets of addresses of the redirected servers.
-	// TODO (manadart 2019-11-08): Change this to be either MachineHostPorts
-	// or the HostPorts indirection. We don't care about space info here.
-	// We can then delete the API params helpers for conversion for this type
-	// as it will no longer be used.
-	Servers []network.ProviderHostPorts `json:"servers"`
-
-	// CACert holds the certificate of the remote server.
-	CACert string `json:"ca-cert"`
-
-	// ControllerTag uniquely identifies the controller being redirected to.
-	ControllerTag names.ControllerTag `json:"controller-tag,omitempty"`
-
-	// An optional alias for the controller where the model got redirected to.
-	ControllerAlias string `json:"controller-alias,omitempty"`
-}
-
-// Error implements the error interface.
-func (e *RedirectError) Error() string {
-	return "redirection to alternative server required"
-}
-
-// IsRedirectError returns true if err is caused by a RedirectError.
-func IsRedirectError(err error) bool {
-	_, ok := errors.Cause(err).(*RedirectError)
-	return ok
-}
 
 var (
 	ErrBadId              = errors.New("id not found")
@@ -277,6 +118,8 @@ func ServerErrorAndStatus(err error) (*params.Error, int) {
 		status = http.StatusUnauthorized
 	case params.CodeRedirect:
 		status = http.StatusMovedPermanently
+	case params.CodeNotLeader:
+		status = http.StatusTemporaryRedirect
 	}
 	return err1, status
 }
@@ -380,6 +223,10 @@ func ServerError(err error) *params.Error {
 		code = params.CodeIncompatibleClient
 		rawErr := errors.Cause(err).(*params.IncompatibleClientError)
 		info = rawErr.AsMap()
+	case IsNotLeaderError(err):
+		code = params.CodeNotLeader
+		rawErr := errors.Cause(err).(*NotLeaderError)
+		info = rawErr.AsMap()
 	default:
 		code = params.ErrCode(err)
 	}
@@ -478,6 +325,20 @@ func RestoreError(err error) error {
 		return err
 	case params.IsCodeQuotaLimitExceeded(err):
 		return errors.NewQuotaLimitExceeded(nil, msg)
+	case params.IsCodeNotLeader(err):
+		e, ok := err.(*params.Error)
+		if !ok {
+			return err
+		}
+		var serverAddress string
+		if s, ok := e.Info["server-address"]; ok {
+			serverAddress = s.(string)
+		}
+		var serverID string
+		if s, ok := e.Info["server-id"]; ok {
+			serverID = s.(string)
+		}
+		return NewNotLeaderError(serverAddress, serverID)
 	default:
 		return err
 	}
